+++ cypress/e2e/log-in.cy.js
@@ -65,7 +65,8 @@ describe("Log in", () => {
 
     cy.url().should("eq", "http://localhost:5173/");
 
-    cy.contains("A list of my favorite books").click();
+    cy.contains("General examples").click();
+    cy.contains("A secret list of my favorite books").click();
     cy.contains("It Ends With Us");
 
     cy.get('[aria-label="Profile"]').click();
@@ -76,15 +77,15 @@ describe("Log in", () => {
 
   it("Do not log in and query on private data unauthenticated", () => {
     cy.visit("/");
-
-    cy.contains("A list of my favorite books").click();
+    cy.contains("General examples").click();
+    cy.contains("A secret list of my favorite books").click();
     cy.contains("It Ends With Us").should("not.exist");
   });
 
   it("Do not log in and query public data", () => {
     cy.visit("/");
-
-    cy.contains("My wish list").click();
+    cy.contains("General examples").click();
+    cy.contains("A public list of books I'd love to own").click();
     cy.contains("Too Late");
   });
 });