+++ cypress/e2e/sources-info.cy.js
@@ -1,8 +1,8 @@
 describe("Sources info", () => {
   it("Fetch status on fetch failure", () => {
     cy.visit("/");
-
-    cy.contains("My favourite musicians").click();
+    cy.contains("General examples").click();
+    cy.contains("A query about musicians").click();
     cy.contains("Finished in:");
     cy.get('[aria-label="Sources info"]').click();
 
@@ -11,8 +11,8 @@ describe("Sources info", () => {
 
   it("Fetch status on fetch success", () => {
     cy.visit("/");
-
-    cy.contains("My wish list").click();
+    cy.contains("General examples").click();
+    cy.contains("A public list of books I'd love to own").click();
     cy.contains("Finished in:");
     cy.get('[aria-label="Sources info"]').click();
 
@@ -21,7 +21,7 @@ describe("Sources info", () => {
 
   it("Fetch status on cached source - see https://github.com/SolidLabResearch/generic-data-viewer-react-admin/issues/59", () => {
     cy.visit("/");
-
+    cy.contains("Project related examples").click();
     cy.contains("Components").click();
     cy.contains("Finished in:");
     cy.get('[aria-label="Sources info"]').click();
@@ -38,8 +38,8 @@ describe("Sources info", () => {
 
   it("Authentication not required for query on public data", () => {
     cy.visit("/");
-
-    cy.contains("My wish list").click();
+    cy.contains("General examples").click();
+    cy.contains("A public list of books I'd love to own").click();
     cy.contains("Finished in:");
     cy.get('[aria-label="Sources info"]').click();
 
@@ -63,7 +63,8 @@ describe("Sources info", () => {
 
     cy.url().should("eq", "http://localhost:5173/");
 
-    cy.contains("A list of my favorite books").click();
+    cy.contains("General examples").click();
+    cy.contains("A secret list of my favorite books").click();
     cy.contains("Finished in:");
     cy.get('[aria-label="Sources info"]').click();
 
@@ -72,8 +73,8 @@ describe("Sources info", () => {
 
   it("Authentication uncertain for query on not existing source", () => {
     cy.visit("/");
-
-    cy.contains("My favourite musicians").click();
+    cy.contains("General examples").click();
+    cy.contains("A query about musicians").click();
     cy.contains("Finished in:");
     cy.get('[aria-label="Sources info"]').click();