+++ cypress/e2e/templated-query.cy.js
@@ -6,7 +6,8 @@ describe("Templated query", () => {
   it("With 1 variable", () => {
 
     cy.visit("/");
-    cy.contains("Templated query about my favourite musicians").click();
+    cy.contains("General examples").click();
+    cy.contains("A templated query about musicians").click();
 
     // Fill in the query, select Baroque (7 existant artists -> perfect for this test)
     cy.get('form').within(() => {
@@ -40,8 +41,8 @@ describe("Templated query", () => {
 
   it("With 2 variables", () => {
     cy.visit("/");
-
-    cy.contains("Templated query #2 about my favourite musicians").click();
+    cy.contains("For testing only").click();
+    cy.contains("A templated query about musicians, two variables").click();
 
     cy.get('form').within(() => {
       cy.get('#genre').click();
@@ -59,7 +60,8 @@ describe("Templated query", () => {
 
   it("Able to change variables after making a templated query", () => {
     cy.visit("/");
-    cy.contains("Templated query about my favourite musicians").click();
+    cy.contains("General examples").click();
+    cy.contains("A templated query about musicians").click();
 
     // Fill in the query
     cy.get('form').within(() => {
@@ -86,7 +88,8 @@ describe("Templated query", () => {
 
   it("Correct message displayed when no resulting data", () => {
     cy.visit("/");
-    cy.contains("Templated query #2 about my favourite musicians").click();
+    cy.contains("For testing only").click();
+    cy.contains("A templated query about musicians, two variables").click();
 
     // Chose a genre
     cy.get('form').within(() => {
@@ -109,7 +112,8 @@ describe("Templated query", () => {
 
   it("Able to change variables after having no results", () => {
     cy.visit("/");
-    cy.contains("Templated query #2 about my favourite musicians").click();
+    cy.contains("For testing only").click();
+    cy.contains("A templated query about musicians, two variables").click();
 
     // Chose a genre
     cy.get('form').within(() => {