+++ cypress/e2e/type-representation.cy.js
@@ -2,7 +2,8 @@ describe("Type representation", () => {
   it("Querying variable ending on _img should return an image as result", () => {
     cy.visit("/");
 
-    cy.contains("A test for images").click();
+    cy.contains("General examples").click();
+    cy.contains("Some images").click();
     cy.contains("Finished in:");
     cy.get("td").find("img");
   });