+++ cypress/e2e/verify-source.cy.js
@@ -1,8 +1,9 @@
 describe("Verify source", () => {
   it("Verification button exists and returns correct result", () => {
     cy.visit("/");
-
-    cy.contains("Components, with source verification").click();
+    
+    cy.contains("General examples").click();
+    cy.contains("Source verification").click();
 
     cy.get('[aria-label="Sources info"]').click();