+++ src/IconProvider/IconProvider.js
@@ -3,11 +3,17 @@ import MenuBookIcon from "@mui/icons-material/MenuBook";
 import MusicNoteIcon from '@mui/icons-material/MusicNote';
 import PhotoIcon from '@mui/icons-material/Photo';
 import FunctionsIcon from '@mui/icons-material/Functions';
+import ListAltIcon from '@mui/icons-material/ListAlt';
+import FactoryIcon from '@mui/icons-material/Factory';
+import BugReportIcon from '@mui/icons-material/BugReport';
 
 export default {
   BrushIcon,
   MenuBookIcon,
   MusicNoteIcon,
   PhotoIcon,
-  FunctionsIcon
+  FunctionsIcon,
+  ListAltIcon,
+  FactoryIcon,
+  BugReportIcon
 };