+++ src/components/InteractionLayout/SelectionMenu/SelectionMenu.jsx
@@ -1,8 +1,18 @@
-import React, { Component } from 'react';
+import React, { Component, useState } from 'react';
 import { useResourceDefinitions } from "ra-core";
 import { DashboardMenuItem } from "ra-ui-materialui";
 import { Menu } from "react-admin";
 import { ThemeProvider, createTheme, Tooltip, Box, Typography } from "@mui/material";
+import config from "../../../config.json";
+import List from '@mui/material/List';
+import ListItemIcon from '@mui/material/ListItemIcon';
+import ListItemButton from '@mui/material/ListItemButton';
+import ListItemText from '@mui/material/ListItemText';
+import Collapse from '@mui/material/Collapse';
+import ExpandLess from '@mui/icons-material/ExpandLess';
+import ExpandMore from '@mui/icons-material/ExpandMore';
+import IconProvider from "../../../IconProvider/IconProvider";
+import ViewListIcon from '@mui/icons-material/ViewList';
 
 /**
  * A custom menu as defined in React Admin for selecting the query the user whishes to execute.
@@ -10,26 +20,70 @@ import { ThemeProvider, createTheme, Tooltip, Box, Typography } from "@mui/mater
  */
 function SelectionMenu() {
   const resources = useResourceDefinitions();
+  const queryGroups = config.queryGroups;
+
+  // adding a list to the group that will contain all the queries for said group
+  queryGroups.forEach(group => group.queries = [])
+
+  // fill in the groups, and put the ones without a group inside the looseQueries list
+  const looseQueries = setUpQueryGroups(queryGroups, resources);
+
   return (
     <ThemeProvider theme={menuItemTheme}>
       <div style={{ height: '100%', overflowY: 'auto', backgroundColor: 'white' }}>
         <Menu>
-          <DashboardMenuItem />
-          {Object.keys(resources).map((id) => (
-            <Tooltip
-              key={id}
-              placement="right"
-              title={
-                <TooltipContent
-                  title={resources[id].options.label}
-                  description={resources[id].options.descr} />
-              }
-            >
-              <div >
-                <Menu.ResourceItem name={id} />
-              </div>
-            </Tooltip>
-          ))}
+          <List>
+            <DashboardMenuItem />
+
+            {looseQueries.map(id => (
+              <Tooltip
+                key={id}
+                placement="right"
+                title={
+                  <TooltipContent
+                    title={resources[id].options.label}
+                    description={resources[id].options.descr} />
+                }
+              >
+                <div >
+                  <Menu.ResourceItem name={id} />
+                </div>
+              </Tooltip>
+            ))}
+          </List>
+          {queryGroups.map((group) => {
+            const [open, setOpen] = useState(false)
+            return (
+              <List key={group.id} disablePadding >
+                <ListItemButton onClick={() => { setOpen(!open) }}>
+                  <ListItemIcon>
+                    {getIconComponent(group.icon)}
+                  </ListItemIcon>
+                  <ListItemText primary={group.name} />
+                  {open ? <ExpandLess /> : <ExpandMore />}
+                </ListItemButton>
+                <Collapse in={open} timeout="auto" unmountOnExit>
+                  <List component="div" disablePadding>
+                    {group.queries.map((id) => (
+                      <Tooltip
+                        key={id}
+                        placement="right"
+                        title={
+                          <TooltipContent
+                            title={resources[id].options.label}
+                            description={resources[id].options.descr} />
+                        }
+                      >
+                        <ListItemText sx={{ overflow: 'hidden', ml: 1.5 }} >
+                          <Menu.ResourceItem name={id} />
+                        </ListItemText>
+                      </Tooltip>
+                    ))}
+                  </List>
+                </Collapse>
+              </List>
+            )
+          })}
         </Menu>
       </div>
     </ThemeProvider>
@@ -61,6 +115,13 @@ const menuItemTheme = createTheme({
   },
 });
 
+const getIconComponent = (iconKey) => {
+  const IconComponent = IconProvider[iconKey];
+  if (IconComponent) {
+    return <IconComponent />;
+  }
+  return <ViewListIcon />;
+};
 
 const TooltipContent = ({ title, description }) => (
   <React.Fragment>
@@ -70,7 +131,6 @@ const TooltipContent = ({ title, description }) => (
         backgroundColor: '#6d6d6d',
         paddingX: 1,
         marginX: -1,
-        
       }}
     >
       <Typography variant="h6" component="div">
@@ -89,5 +149,20 @@ const TooltipContent = ({ title, description }) => (
   </React.Fragment>
 )
 
+const setUpQueryGroups = (queryGroups, resources) => {
+  const looseQueries = [];
+  Object.keys(resources).forEach((id) => {
+    try {
+      if (resources[id].options.queryGroupId === undefined) {
+        looseQueries.push(id)
+      } else {
+        queryGroups.find(group => group.id === resources[id].options.queryGroupId).queries.push(id)
+      }
+    } catch (error) {
+      throw new Error(`Error adding queries to a group: ${error.message}`);
+    }
+  })
+  return looseQueries;
+}
 
 export default SelectionMenu;